flytter FlereArbeidsforhold til egen fil
